Clean up linter mess.
