+++ src/touchosc/mod.rs
@@ -5,17 +5,9 @@ mod serialize;
 pub mod serve;
 mod templates;
 
-#[expect(unused)]
-pub use generate::generate_layout;
 pub use generate::{GroupEntry, assemble_layout};
 pub use model::*;
-#[expect(unused)]
-pub use parse::parse_touchosc;
-#[expect(unused)]
-pub use templates::BASE_TEMPLATE;
-#[expect(unused)]
-pub use templates::load_base_template;
-#[expect(unused)]
+#[cfg(test)]
 pub use templates::load_group_template;
 pub use templates::{TEMPLATES, TemplateEntry};
